Moved plot into lib files Draw base line, min and max on plot

+++ Lib/s_string.c
@@ -64,9 +64,9 @@ s8 s8_f64_to_s8(Arena* arena, f64 num, isize presicion)
 
   s8_append(arena, &res, s8_isize_to_s8(arena, (isize)int_part));
 
-  frac_part = frac_part * pow(10, presicion);
+  frac_part = frac_part * pow(10, (f64)presicion);
 
-  s8 frac_s = s8_isize_to_s8(arena, frac_part);
+  s8 frac_s = s8_isize_to_s8(arena, (isize)frac_part);
 
   s8_append(arena, &res, s8_from_literal("."));
   s8_append(arena, &res, frac_s);
